requirements.txt
# requirements.txt
pillow==10.2.0
python-barcode==0.15.1
pywin32==306
setuptools==69.1.1
six==1.16.0
wxPython==4.2.1
